Move room construction logic into factory class
