+++ app/src/community/java/com/twilio/video/app/VideoApplicationComponent.java
@@ -18,7 +18,7 @@ package com.twilio.video.app;
 
 import com.twilio.video.app.auth.CommunityAuthModule;
 import com.twilio.video.app.data.AuthServiceModule;
-import com.twilio.video.app.sdk.CommunityRoomManagerModule;
+import com.twilio.video.app.sdk.CommunityVideoSdkModule;
 import com.twilio.video.app.security.SecurityModule;
 import com.twilio.video.app.ui.CommunityScreenSelectorModule;
 import com.twilio.video.app.ui.login.CommunityLoginActivityModule;
@@ -45,7 +45,7 @@ import dagger.android.AndroidInjectionModule;
         SettingsActivityModule.class,
         SettingsFragmentModule.class,
         VideoServiceModule.class,
-        CommunityRoomManagerModule.class,
+        CommunityVideoSdkModule.class,
         SecurityModule.class,
         AudioSwitchModule.class
     }
