+++ app/src/main/java/com/twilio/video/app/sdk/VideoSdkModule.kt
@@ -4,25 +4,31 @@ import android.app.Application
 import android.content.SharedPreferences
 import com.twilio.video.app.ApplicationModule
 import com.twilio.video.app.ApplicationScope
-import com.twilio.video.app.data.AuthServiceModule
 import com.twilio.video.app.data.DataModule
 import com.twilio.video.app.data.api.TokenService
+import com.twilio.video.app.data.api.VideoAppServiceModule
 import dagger.Module
 import dagger.Provides
 
 @Module(includes = [
     ApplicationModule::class,
     DataModule::class,
-    AuthServiceModule::class])
-class CommunityRoomManagerModule {
+    VideoAppServiceModule::class])
+class VideoSdkModule {
+
+    @Provides
+    fun providesRoomFactory(
+        application: Application,
+        sharedPreferences: SharedPreferences,
+        tokenService: TokenService
+    ): RoomFactory =
+            RoomFactory(application, sharedPreferences, tokenService)
 
     @Provides
     @ApplicationScope
     fun providesRoomManager(
         application: Application,
-        sharedPreferences: SharedPreferences,
-        tokenService: TokenService
-    ): RoomManager {
-        return RoomManager(application, sharedPreferences, tokenService)
-    }
+        roomFactory: RoomFactory
+    ): RoomManager =
+            RoomManager(application, roomFactory)
 }
