+++ app/src/twilio/java/com/twilio/video/app/VideoApplicationComponent.java
@@ -3,7 +3,7 @@ package com.twilio.video.app;
 import com.twilio.video.app.auth.AuthModule;
 import com.twilio.video.app.data.DataModule;
 import com.twilio.video.app.data.api.VideoAppServiceModule;
-import com.twilio.video.app.sdk.RoomManagerModule;
+import com.twilio.video.app.sdk.VideoSdkModule;
 import com.twilio.video.app.ui.ScreenSelectorModule;
 import com.twilio.video.app.ui.login.LoginActivityModule;
 import com.twilio.video.app.ui.room.RoomActivityModule;
@@ -30,7 +30,7 @@ import dagger.android.AndroidInjectionModule;
         SettingsActivityModule.class,
         SettingsFragmentModule.class,
         VideoServiceModule.class,
-        RoomManagerModule.class,
+        VideoSdkModule.class,
         AudioSwitchModule.class
     }
 )
